preliminary sticky support on scrollable DIVs
I like this implementation, but it lacked the ability for me to have sitcky elements on DIV containers that were scrollable. In my app the scrolling sections are not on the entire BODY but rather predefined viewports that have their `overflow` turned on.

Added two attributes that could be supplied to the directive:

`scrolling-parent-selector`
A selector-style syntax that scans up the DOM tree to finding the DIV that is responsible for the scrolling. This is nice loose coupling. Don't want to assume its just the parenNpode.

`is-scroll-height-observed`
This flag is needed in the case where the scrollable content changes its content, which in turn triggers a change in the total scrollable height of the above mentioned scrolling viewport. For example I have an expandable tree inside. I was sure which calculations need to be performed, so I reused the one from onResize ( of the web browser ). Hopefully that is "good enough"

I didn't do comprehensive testing, perhaps some others can take this idea further?