exposed another public method stickResetLayout
This allows content management rules to trigger a reset of the sticky position and state if the content in the sticky "resets" ( or changes drastically ). This would be beyond the normal scope of the Sticky directive to be concerned about.

Is there a better way to expose a few public methods to invoke from outside controllers?? As long as they are on the same scope chain the way I did using $parent should be fine right?

In my use case, the content inside the sticky ( which is a title ) changes. The changes are significant enough to mess up rendering severely unless some resets are done in between the content updates.