Update sticky.js
the suggested updates are to fix an issue where the object unstick itself and then resticks itself. it was not calculating the left position correctly.  This has it calculate its position from the left side of the browser.  Which is where the fixed element is probably going to calculate from.