fix(onChange fn): ensure negative value of sticky element offset  does not trigger setting up a new "sticky line"
This fixes the stuck element's behaviour when used together with angular-material's $mdDialog service. When a dialog is shown, angular-material changes "position" and "top" properties of the "body" element for the time when popup is visible. This brakes "onChange" fn's calculations of the sticky line which tells ngStickhy when to "un-stick" the sticky element when user scrolls back up.

fix #130